Search相关组件改写为tsx
