+++ src/renderer/src/components/search/lib/Header.tsx
@@ -1,16 +1,19 @@
-<template>
+
+interface InputInterface {
+    input: string
+}
+
+const Header = ({input}: InputInterface) => (
     <div>
         <header class="flex flex-row items-end justify-start gap-2">
             <h1>Search</h1>
             <h1 class="underline underline-offset-1">
-                <mark>{{ props.input }}</mark>
+                <mark>{ input }</mark>
             </h1>
         </header>
     </div>
-</template>
+)
 
-<script lang="ts" setup>
-const props = defineProps<{
-    input: string
-}>()
-</script>
+export {
+    Header
+}
